2021/python/aoc2021/eight/tests/test_eight.py
# tests for day eight

from ..eight import parse_input, part_one_calc, part_two_calc, segs_to_num

input = """
be cfbegad cbdgef fgaecd cgeb fdcge agebfd fecdb fabcd edb | fdgacbe cefdb cefbgd gcbe
edbfga begcd cbg gc gcadebf fbgde acbgfd abcde gfcbed gfec | fcgedb cgb dgebacf gc
fgaebd cg bdaec gdafb agbcfd gdcbef bgcad gfac gcb cdgabef | cg cg fdcagb cbg
fbegcd cbd adcefb dageb afcb bc aefdc ecdab fgdeca fcdbega | efabcd cedba gadfec cb
aecbfdg fbg gf bafeg dbefa fcge gcbea fcaegb dgceab fcbdga | gecf egdcabf bgf bfgea
fgeab ca afcebg bdacfeg cfaedg gcfdb baec bfadeg bafgc acf | gebdcfa ecba ca fadegcb
dbcfg fgd bdegcaf fgec aegbdf ecdfab fbedc dacgb gdcebf gf | cefg dcbef fcge gbcadfe
bdfegc cbegaf gecbf dfcage bdacg ed bedf ced adcbefg gebcd | ed bcgafe cdgba cbgef
egadfb cdbfeg cegd fecab cgb gbdefca cg fgcdab egfdb bfceg | gbdfcae bgc cg cgb
gcafb gcf dcaebfg ecagb gf abcdeg gaef cafbge fdbac fegbdc | fgae cfgab fg bagce
"""


def test_parse_input() -> None:
    data = parse_input(input)
    assert len(data) == 10


def test_part_one() -> None:
    data = parse_input(input)
    assert part_one_calc(data) == 26


def test_digits() -> None:
    assert segs_to_num(["bcdf", "abcefg", "abcdfg", "abdefg"]) == 4096
    assert segs_to_num(["cf"]) == 1


def test_part_two() -> None:
    data = parse_input(input)
    assert part_two_calc(data) == 61229
